chore: add spinner for deleting

+++ frontend/src/database/db.ts
@@ -102,13 +102,15 @@ const deleteDataForPathogen = async (pathogen_id: number) => {
     const distanceMatrix = await getDistanceMatrixByPathogenId(pathogen_id);
     const cases = await db.cases.where({ pathogen_id: pathogen_id }).toArray();
     for (const caseData of cases) {
-        db.samples.where({ fasta_id: caseData.fasta_id }).delete();
-        db.contacts.where({ case_id_1: caseData.id }).or("case_id_2").equals(caseData.id).delete();
-        db.cases.where({ id: caseData.id }).delete();
+        if (caseData.fasta_id) {
+            await db.samples.where({ fasta_id: caseData.fasta_id }).delete();
+        }
+        await db.contacts.where({ case_id_1: caseData.id }).or("case_id_2").equals(caseData.id).delete();
+        await db.cases.where({ id: caseData.id }).delete();
     }
     if (distanceMatrix?.id) {
-        db.distances.where({ distance_matrix_id: distanceMatrix.id }).delete();
-        db.distance_matrices.where({ id: distanceMatrix.id }).delete();
+        await db.distances.where({ distance_matrix_id: distanceMatrix.id }).delete();
+        await db.distance_matrices.where({ id: distanceMatrix.id }).delete();
     }
 };
